+General Refactoring Moved ComponentWithRouter into general component directory Moved updateAction into root directory Removed unused coded on forms Removed redundant functions in Navbar (log wrappers for auth functions) Fixed bug for Reseed users having "createdBy" field be "user" instead of "seed"
