+++ src/components/Login.js
@@ -18,7 +18,7 @@ export function Login() {
 
   return (
     <div className="max-w-[600px] text-center my-2.5 mx-auto pa-[15px] font-sans">
-      <h1 className="text-white text-3xl font-bold pb-3 border-b">Login</h1>
+      <h1 className="pb-3 text-3xl font-bold text-white border-b">Login</h1>
       <label className="leading-loose text-left block mb-3.5 mt-5 text-white text-sm font-extralight">
         Username:{' '}
         <input
@@ -32,14 +32,14 @@ export function Login() {
       <label className="leading-loose text-left block mb-3.5 mt-5 text-white text-sm font-extralight">
         Password:{' '}
         <input
-          className='"text-black inline-block box-border w-full border rounded border-white px-3.5 py-2.5 mb-2.5 text-lg'
+          className="text-black inline-block box-border w-full border rounded border-white px-3.5 py-2.5 mb-2.5 text-lg"
           type="password"
           name="passwordInput"
           onChange={setLoginInput}
         />
         {/* <input type="text" onChange={(e) => setPassword(e.target.value)} /> */}
         <button
-          className="bg-pink-500 text-white uppercase mt-10 p-5 text-base font-thin tracking-wide inline-block appearance-none border-none rounded w-full"
+          className="inline-block w-full p-5 mt-10 text-base font-thin tracking-wide text-white uppercase bg-pink-500 border-none rounded appearance-none"
           type="submit"
           onClick={handleLogin}
         >
